Remove duplicate ForUtil
We copied this into our codebase twice. The version with customizable
block size is unnecessary, we always use the same block size anyway.
=> no need to duplicate 1k lines.
